// app/models/types.tsx
export type JobData = {
  id: number;
  company: string;
  logo: string;
  logoBackground: string;
  position: string;
  postedAt: string;
  contract: string;
  location: string;
  website: string;
  apply: string;
  description: string;
  requirements: {
    content: string;
    items: [];
  };
  role: {
    content: string;
    items: [];
  };
};